added additional components
