+++ src/components/Header.js
@@ -1,0 +1,17 @@
+import React, { Component } from 'react';
+import logo from '../App-Pinterest-icon.png';
+
+class Header extends Component {
+  render() {
+    return (
+      <div className="Header">
+        <header className="App-header">
+          <img src={logo} className="App-logo" alt="logo" />
+          <h1 className="App-title">A Pinterest Simulation Prototype</h1>
+        </header>
+      </div>
+    );
+  }
+}
+
+export default Header;
